+++ src/Edge2D.h
@@ -17,36 +17,36 @@ public:
     virtual bool intercept(const Point2D &p1, const Point2D &p2) const;
     virtual bool intercept(const Point2D *p1, const Point2D *p2) const;
 
-    virtual float distance(const Point2D &p) const;
-    virtual float distance(const Point2D *p) const;
-    virtual float distance(float x, float y) const;
+    virtual double distance(const Point2D &p) const;
+    virtual double distance(const Point2D *p) const;
+    virtual double distance(double x, double y) const;
 
-    virtual float straightDistance(const Point2D &p) const;
-    virtual float straightDistance(const Point2D *p) const;
-    virtual float straightDistance(float x, float y) const;
+    virtual double straightDistance(const Point2D &p) const;
+    virtual double straightDistance(const Point2D *p) const;
+    virtual double straightDistance(double x, double y) const;
 
-    virtual float height(const Point2D &p) const;
-    virtual float height(const Point2D *p) const;
-    virtual float height(float x, float y) const;
+    virtual double height(const Point2D &p) const;
+    virtual double height(const Point2D *p) const;
+    virtual double height(double x, double y) const;
 
     virtual bool left(const Point2D &p) const;
     virtual bool left(const Point2D *p) const;
-    virtual bool left(float x, float y) const;
+    virtual bool left(double x, double y) const;
 
     virtual bool right(const Point2D &p) const;
     virtual bool right(const Point2D *p) const;
-    virtual bool right(float x, float y) const;
+    virtual bool right(double x, double y) const;
 
     virtual bool accordingToNormal(const Point2D &p, bool insideTest = false) const;
 
-    virtual float angle(const Point2D &p1, const Point2D &p2) const;
-    virtual float angle(const Point2D *p1, const Point2D *p2) const;
-    virtual float angle(const Edge2D &e) const;
-    virtual float angle(const Edge2D *e) const;
-    virtual float angle(const Point2D &p) const;
-    virtual float angle(const Point2D *p) const;
-    virtual float orientedAngle(const Point2D &p, bool insideTest = false) const;
-    virtual float orientedAngle(const Point2D *p, bool insideTest = false) const;
+    virtual double angle(const Point2D &p1, const Point2D &p2) const;
+    virtual double angle(const Point2D *p1, const Point2D *p2) const;
+    virtual double angle(const Edge2D &e) const;
+    virtual double angle(const Edge2D *e) const;
+    virtual double angle(const Point2D &p) const;
+    virtual double angle(const Point2D *p) const;
+    virtual double orientedAngle(const Point2D &p, bool insideTest = false) const;
+    virtual double orientedAngle(const Point2D *p, bool insideTest = false) const;
 
     virtual Vector2D normal() const;
