+++ src/Point.h
@@ -12,7 +12,7 @@ class Point
 {
 public:
     Point(long int id = 0);
-    Point(std::vector<float> coords, long int id = 0);
+    Point(std::vector<double> coords, long int id = 0);
 
     Point(const Point *p);
     Point(const Vector &v);
@@ -24,12 +24,12 @@ public:
 
     virtual unsigned int dimension() const;
 
-    virtual void setCoords(std::vector<float> coords);
+    virtual void setCoords(std::vector<double> coords);
 
-    virtual std::vector<float> getCoords() const;
+    virtual std::vector<double> getCoords() const;
 
-    virtual void setCoord(unsigned int i, float c);
-    virtual float getCoord(unsigned int i) const;
+    virtual void setCoord(unsigned int i, double c);
+    virtual double getCoord(unsigned int i) const;
 
     virtual bool match(const Point &p) const;
     virtual bool match(const Point *p) const;
@@ -37,20 +37,20 @@ public:
     virtual bool equal(const Point &p) const;
     virtual bool equal(const Point *p) const;
 
-    float distance(const Point &p) const;
+    double distance(const Point &p) const;
 
-    void setColorR(float color) {r = color;};
-    void setColorG(float color) {g = color;};
-    void setColorB(float color) {b = color;};
+    void setColorR(double color) {r = color;};
+    void setColorG(double color) {g = color;};
+    void setColorB(double color) {b = color;};
 
     virtual void draw() const;
 
 protected:
     int id;
-    float r;
-    float g;
-    float b;
-    std::vector<float> coords;
+    double r;
+    double g;
+    double b;
+    std::vector<double> coords;
 
 
 };
