+++ src/Point2D.cpp
@@ -1,7 +1,7 @@
 #include "Point2D.h"
 #include <GL/gl.h>
 
-Point2D::Point2D(float x, float y, long int id) :
+Point2D::Point2D(double x, double y, long int id) :
 	Point(id)
 {
     this->coords.resize(2, 0.0);
@@ -35,23 +35,23 @@ unsigned int Point2D::dimension() const
 	return 2;
 }
 
-void Point2D::setPosition(float x, float y)
+void Point2D::setPosition(double x, double y)
 {
 	this->setX(x);
 	this->setY(y);
 }
 
-float Point2D::surface(const Point2D &p1, const Point2D &p2) const
+double Point2D::surface(const Point2D &p1, const Point2D &p2) const
 {
 	return fabs(this->orientedSurface(p1, p2));
 }
 
-float Point2D::surface(const Point2D *p1, const Point2D *p2) const
+double Point2D::surface(const Point2D *p1, const Point2D *p2) const
 {
 	return this->surface(*p1, *p2);
 }
 
-float Point2D::orientedSurface(const Point2D &p1, const Point2D &p2) const
+double Point2D::orientedSurface(const Point2D &p1, const Point2D &p2) const
 {
 	Vector2D ab(*this, p1);
 	Vector2D ac(*this, p2);
@@ -59,32 +59,32 @@ float Point2D::orientedSurface(const Point2D &p1, const Point2D &p2) const
 	return 0.5*ab.cross(ac);
 }
 
-float Point2D::orientedSurface(const Point2D *p1, const Point2D *p2) const
+double Point2D::orientedSurface(const Point2D *p1, const Point2D *p2) const
 {
 	return this->orientedSurface(*p1, *p2);
 }
 
-void Point2D::setX(float x)
+void Point2D::setX(double x)
 {
 	this->setCoord(0, x);
 }
 
-float Point2D::getX() const
+double Point2D::getX() const
 {
 	return this->getCoord(0);
 }
 
-void Point2D::setY(float y)
+void Point2D::setY(double y)
 {
 	this->setCoord(1, y);
 }
 
-float Point2D::getY() const
+double Point2D::getY() const
 {
 	return this->getCoord(1);
 }
 
-bool Point2D::equal(float x, float y) const
+bool Point2D::equal(double x, double y) const
 {
 	Point2D p(x, y);
