+++ src/Vector2D.h
@@ -10,31 +10,31 @@ class Vector2D : public Vector
 {
 public:
 
-	Vector2D(float x = 0.0, float y = 0.0, long int id = 0);
+	Vector2D(double x = 0.0, double y = 0.0, long int id = 0);
 	Vector2D(const Point2D &p1, const Point2D &p2);
 	Vector2D(const Vector &v);
 	Vector2D(const Vector *v);
 	virtual ~Vector2D();
 
-	virtual void setPosition(float x, float y);
+	virtual void setPosition(double x, double y);
 
-	virtual void setX(float x);
-	virtual float getX() const;
+	virtual void setX(double x);
+	virtual double getX() const;
 
-	virtual void setY(float y);
-	virtual float getY() const;
+	virtual void setY(double y);
+	virtual double getY() const;
 
 	virtual unsigned int dimension() const;
 
-	virtual float cross(const Vector2D &v) const;
-	virtual float cross(const Vector2D *v) const;
+	virtual double cross(const Vector2D &v) const;
+	virtual double cross(const Vector2D *v) const;
 
-	virtual float scaledCross(const Vector2D &v) const;
-	virtual float scaledCross(const Vector2D *v) const;
+	virtual double scaledCross(const Vector2D &v) const;
+	virtual double scaledCross(const Vector2D *v) const;
 
-	virtual float orientedAngle() const;
-	virtual float orientedAngle(const Vector2D &v) const;
-	virtual float orientedAngle(const Vector2D *v) const;
+	virtual double orientedAngle() const;
+	virtual double orientedAngle(const Vector2D &v) const;
+	virtual double orientedAngle(const Vector2D *v) const;
 
 	virtual bool left(const Vector2D &v) const;
 	virtual bool left(const Vector2D *v) const;
